[FE] REFACTOR: 메뉴 훅 생성(useMenu)을 통한 리팩토링
